+++ admin/src/components/map/layers.ts
@@ -6,7 +6,7 @@ import { Fill } from 'ol/style';
 import Map from 'ol/Map';
 import Feature, { FeatureLike } from 'ol/Feature';
 import { getMap } from './DvkMap';
-import { FairwayCardPartsFragment, HarborPartsFragment, Orientation, Quay, Section } from '../../graphql/generated';
+import { HarborPartsFragment, Orientation, Quay, Section } from '../../graphql/generated';
 import { FeatureDataLayerId, FeatureLayerId, MAP } from '../../utils/constants';
 import * as olExtent from 'ol/extent';
 import { getFairwayArea12Style } from './layerStyles/fairwayArea12Styles';
@@ -19,7 +19,6 @@ import { getSafetyEquipmentStyle } from './layerStyles/safetyEquipmentStyles';
 import { GeoJSON } from 'ol/format';
 import { getVtsStyle } from './layerStyles/vtsStyles';
 import { getCircleStyle } from './layerStyles/circleStyles';
-import { getFairwayAreaBorderFeatures } from '../../fairwayareaworker/FairwayAreaUtils';
 import { Geometry } from 'ol/geom';
 import { getPilotRouteStyle } from './layerStyles/pilotRouteStyles';
 import { getPilotageLimitStyle } from './layerStyles/pilotageLimitStyles';
@@ -380,7 +379,7 @@ export function addAPILayers(map: Map) {
   });
 }
 
-function getFittingPadding() {
+export function getFittingPadding() {
   const dvkMap = getMap();
   const fitPadding = [100, 50, 50, 50];
   const size = dvkMap.olMap?.getSize() ?? [0, 0];
@@ -416,62 +415,6 @@ export function fitSelectedFairwayCardOnMap() {
   if (!olExtent.isEmpty(extent)) {
     dvkMap.olMap?.getView().fit(extent, { padding: getFittingPadding(), duration: 1000 });
   }
-}
-
-export function unsetSelectedFairwayCard() {
-  const dvkMap = getMap();
-  const line12Source = dvkMap.getVectorSource('line12');
-  const line3456Source = dvkMap.getVectorSource('line3456');
-  const area12Source = dvkMap.getVectorSource('area12');
-  const area3456Source = dvkMap.getVectorSource('area3456');
-  const quaySource = dvkMap.getVectorSource('quay');
-  const selectedFairwayCardSource = dvkMap.getVectorSource('selectedfairwaycard');
-  const depthSource = dvkMap.getVectorSource('depth12');
-  const specialArea2Source = dvkMap.getVectorSource('specialarea2');
-  const specialArea15Source = dvkMap.getVectorSource('specialarea15');
-  const boardLine12Source = dvkMap.getVectorSource('boardline12');
-  const harborSource = dvkMap.getVectorSource('harbor');
-  const circleSource = dvkMap.getVectorSource('circle');
-  const oldSelectedFeatures = selectedFairwayCardSource.getFeatures().concat(quaySource.getFeatures());
-  for (const feature of oldSelectedFeatures) {
-    switch (feature.getProperties().dataSource) {
-      case 'line12':
-        line12Source.addFeature(feature);
-        feature.unset('n2000HeightSystem');
-        break;
-      case 'line3456':
-        line3456Source.addFeature(feature);
-        break;
-      case 'area12':
-        area12Source.addFeature(feature);
-        feature.unset('n2000HeightSystem');
-        (depthSource.getFeatureById(feature.getId() as number) as Feature<Geometry>)?.unset('n2000HeightSystem');
-        break;
-      case 'area3456':
-        area3456Source.addFeature(feature);
-        break;
-      case 'specialarea2':
-        specialArea2Source.addFeature(feature);
-        feature.unset('n2000HeightSystem');
-        break;
-      case 'specialarea15':
-        specialArea15Source.addFeature(feature);
-        feature.unset('n2000HeightSystem');
-        break;
-      case 'boardline12':
-        boardLine12Source.addFeature(feature);
-        break;
-      case 'harbor':
-        harborSource.addFeature(feature);
-        break;
-      case 'circle':
-        circleSource.addFeature(feature);
-        break;
-    }
-  }
-  selectedFairwayCardSource.getFeatures().forEach((f) => f.set('selected', false, true));
-  selectedFairwayCardSource.clear();
-  quaySource.clear();
 }
 
 function addQuayFeature(harbor: HarborPartsFragment, quay: Quay, features: VectorSource, format: GeoJSON, showDepth: boolean) {
@@ -513,7 +456,7 @@ function addSectionFeature(harbor: HarborPartsFragment, quay: Quay, section: Sec
   features.addFeature(feature);
 }
 
-function addQuay(harbor: HarborPartsFragment, source: VectorSource) {
+export function addQuay(harbor: HarborPartsFragment, source: VectorSource) {
   const format = new GeoJSON();
   for (const quay of harbor.quays ?? []) {
     let sectionGeometryMissing = false;
@@ -528,119 +471,6 @@ function addQuay(harbor: HarborPartsFragment, source: VectorSource) {
     });
     if (quay?.geometry) {
       addQuayFeature(harbor, quay, source, format, sectionGeometryMissing);
-    }
-  }
-}
-
-export function setSelectedFairwayCard(fairwayCard: FairwayCardPartsFragment | undefined) {
-  const dvkMap = getMap();
-  if (fairwayCard) {
-    const line12Source = dvkMap.getVectorSource('line12');
-    const line3456Source = dvkMap.getVectorSource('line3456');
-    const area12Source = dvkMap.getVectorSource('area12');
-    const area3456Source = dvkMap.getVectorSource('area3456');
-    const quaySource = dvkMap.getVectorSource('quay');
-    const selectedFairwayCardSource = dvkMap.getVectorSource('selectedfairwaycard');
-    const depthSource = dvkMap.getVectorSource('depth12');
-    const specialArea2Source = dvkMap.getVectorSource('specialarea2');
-    const specialArea15Source = dvkMap.getVectorSource('specialarea15');
-    const boardLine12Source = dvkMap.getVectorSource('boardline12');
-    const harborSource = dvkMap.getVectorSource('harbor');
-    const circleSource = dvkMap.getVectorSource('circle');
-    unsetSelectedFairwayCard();
-
-    const fairwayFeatures: Feature[] = [];
-    for (const fairway of fairwayCard?.fairways || []) {
-      for (const line of fairway.navigationLines ?? []) {
-        let feature = line12Source.getFeatureById(line.id) as Feature<Geometry>;
-        if (feature) {
-          line12Source.removeFeature(feature);
-          fairwayFeatures.push(feature);
-          feature.set('n2000HeightSystem', fairwayCard?.n2000HeightSystem || false);
-        } else {
-          feature = line3456Source.getFeatureById(line.id) as Feature<Geometry>;
-          if (feature) {
-            line3456Source.removeFeature(feature);
-            fairwayFeatures.push(feature);
-          }
-        }
-      }
-      for (const area of fairway.areas ?? []) {
-        let feature = area12Source.getFeatureById(area.id) as Feature<Geometry>;
-        if (feature) {
-          area12Source.removeFeature(feature);
-          fairwayFeatures.push(feature);
-          feature.set('n2000HeightSystem', fairwayCard?.n2000HeightSystem || false);
-          feature = depthSource.getFeatureById(area.id) as Feature<Geometry>;
-          feature?.set('n2000HeightSystem', fairwayCard?.n2000HeightSystem || false);
-        } else {
-          feature = area3456Source.getFeatureById(area.id) as Feature<Geometry>;
-          if (feature) {
-            area3456Source.removeFeature(feature);
-            fairwayFeatures.push(feature);
-          }
-        }
-        if (!feature) {
-          feature = specialArea2Source.getFeatureById(area.id) as Feature<Geometry>;
-          if (feature) {
-            specialArea2Source.removeFeature(feature);
-            fairwayFeatures.push(feature);
-            feature.set('n2000HeightSystem', fairwayCard?.n2000HeightSystem || false);
-          }
-        }
-      }
-      for (const prohibitionArea of fairway.prohibitionAreas ?? []) {
-        const feature = specialArea15Source.getFeatureById(prohibitionArea.id) as Feature<Geometry>;
-        if (feature) {
-          specialArea15Source.removeFeature(feature);
-          fairwayFeatures.push(feature);
-        }
-      }
-      for (const line of fairway.boardLines ?? []) {
-        const feature = boardLine12Source.getFeatureById(line.id) as Feature<Geometry>;
-        if (feature) {
-          boardLine12Source.removeFeature(feature);
-          fairwayFeatures.push(feature);
-        }
-      }
-      for (const circle of fairway.turningCircles ?? []) {
-        const feature = circleSource.getFeatureById(circle.id) as Feature<Geometry>;
-        if (feature) {
-          circleSource.removeFeature(feature);
-          fairwayFeatures.push(feature);
-        }
-      }
-    }
-
-    for (const harbor of fairwayCard?.harbors ?? []) {
-      const id = harbor.geometry?.coordinates?.join(';');
-      const feature = id ? (harborSource.getFeatureById(id) as Feature<Geometry>) : undefined;
-      if (feature) {
-        harborSource.removeFeature(feature);
-        fairwayFeatures.push(feature);
-      }
-      addQuay(harbor, quaySource);
-    }
-
-    const area12Features = fairwayFeatures.filter((f) => f.get('dataSource') === 'area12');
-    const borderLineFeatures = getFairwayAreaBorderFeatures(area12Features);
-    borderLineFeatures.forEach((f) => {
-      f.set('dataSource', 'area12Borderline', true);
-      fairwayFeatures.push(f);
-    });
-
-    fairwayFeatures.forEach((f) => f.set('selected', true, true));
-    selectedFairwayCardSource.addFeatures(fairwayFeatures);
-
-    const extent = olExtent.createEmpty();
-    for (const feature of fairwayFeatures) {
-      const geom = feature.getGeometry();
-      if (geom) {
-        olExtent.extend(extent, geom.getExtent());
-      }
-    }
-    if (!olExtent.isEmpty(extent)) {
-      dvkMap.olMap?.getView().fit(extent, { padding: getFittingPadding(), duration: 1000 });
     }
   }
 }
