+++ admin/src/components/pictures/MapExportTool.tsx
@@ -26,7 +26,7 @@ import {
   useVtsPointLayer,
 } from '../map/FeatureLoader';
 import { Fairway, FairwayCardInput, Harbor, Orientation, PictureInput, PictureUploadInput } from '../../graphql/generated';
-import { setSelectedFairwayCard } from '../map/layers';
+import { setSelectedFairwayCard } from '../map/FairwayCardSetter';
 import { useIsFetching } from '@tanstack/react-query';
 import './MapExportTool.css';
 import { useTranslation } from 'react-i18next';
